move Close to separate file
