+++ core/src/commonTest/kotlin/dev/lennartegb/vec2compose/core/PathParserTest.kt
@@ -1,5 +1,6 @@
 package dev.lennartegb.vec2compose.core
 
+import dev.lennartegb.vec2compose.core.commands.Close
 import dev.lennartegb.vec2compose.core.commands.Command
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -30,7 +31,7 @@ internal class PathParserTest {
                 Command.MoveTo(x = 27.05f, y = 24.55f, isAbsolute = true),
                 Command.LineTo(x = 12.15f, y = 19.65f, isAbsolute = true),
                 Command.CurveTo(11.85f, 19.35f, 11.483f, 19.2f, 11.05f, 19.2f, isAbsolute = true),
-                Command.Close
+                Close
             )
         )
     }
