+++ svg/src/commonMain/kotlin/dev/lennartegb/vec2compose/svg/SVGParser.kt
@@ -9,7 +9,7 @@ import dev.lennartegb.vec2compose.core.ImageVectorParser
 import dev.lennartegb.vec2compose.core.Scale
 import dev.lennartegb.vec2compose.core.Translation
 import dev.lennartegb.vec2compose.core.commands.ArcTo
-import dev.lennartegb.vec2compose.core.commands.Command.Close
+import dev.lennartegb.vec2compose.core.commands.Close
 import dev.lennartegb.vec2compose.core.commands.Command.LineTo
 import dev.lennartegb.vec2compose.core.commands.Command.MoveTo
 import dev.lennartegb.vec2compose.core.parsePath
