+++ svg/src/commonTest/kotlin/dev/lennartegb/vec2compose/svg/SVGParserTest.kt
@@ -6,8 +6,7 @@ import dev.lennartegb.vec2compose.core.ImageVector.Path.Stroke
 import dev.lennartegb.vec2compose.core.Scale
 import dev.lennartegb.vec2compose.core.Translation
 import dev.lennartegb.vec2compose.core.commands.ArcTo
-import dev.lennartegb.vec2compose.core.commands.Command
-import dev.lennartegb.vec2compose.core.commands.Command.Close
+import dev.lennartegb.vec2compose.core.commands.Close
 import dev.lennartegb.vec2compose.core.commands.Command.MoveTo
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -167,7 +166,7 @@ internal class SVGParserTest {
                                 x1 = 0f,
                                 y1 = 20f
                             ),
-                            Command.Close
+                            Close
                         )
                     )
                 )
